+++ CHANGELOG.md
@@ -4,6 +4,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.2.1] - 2025-11-05
+
+### Versions
+
+- CSC FastAPI Backend:  0.4.0.4
+- CSC React Frontend:   0.4.0.5
+- CSC Grasshopper Interface: 0.4.2.0
+
+### Added
+
+#### CSC React Frontend
+
+- **Added direct UUID lookup**: On the browse page you can now enter a UUID directly and go to its detail page
+
 ## [0.4.2.0] - 2025-11-02
 
 ### Versions